bot/requirements.txt
discord.py
mysql-connector-python==8.0.22
pandas==1.1.4
praw==7.1.0
python-dateutil==2.8.1
numpy==1.19.3
